i#6262: fixed type error
Issue: #6262

+++ clients/drcachesim/scheduler/scheduler.cpp
@@ -840,7 +840,7 @@ scheduler_tmpl_t<RecordType, ReaderType>::read_traced_schedule()
             }
             cur_cpu = entry.cpu;
             VPRINT(this, 1, "Output #%d is CPU #%" PRId64 "\n", cur_output, cur_cpu);
-            outputs_[cur_output].cpu = cur_cpu;
+            outputs_[cur_output].cpu = (int)cur_cpu;
         }
         input_ordinal_t input = tid2input[entry.thread];
         // We'll fill in the stop ordinal in our second pass below.
